application/pyproject.toml
[tool.poetry]
authors = ["shun198"]
description = "api"
name = "api"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Authlib = "^1.1.0"
Django = "^4.1.2"
boto3 = "^1.26.22"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.24.2"
gunicorn = "^20.1.0"
injector = "^0.20.1"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
pydantic = "^1.10.6"
python = "^3.11"
qrcode = "^7.4.2"
redis = "^4.3.4"
django-rest-knox = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
factory-boy = "^3.2.1"
isort = "^5.11.4"
pdoc = "^13.0.0"
pytest = "^7.1.3"
pytest-bdd = "^6.1.1"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0.0"
pytest-custom-exit-code = "^0.3.0"
pytest-django = "^4.5.2"
pytest-github-actions-annotate-failures = "^0.2.0"
pytest-mock = "^3.10.0"
pytest-sugar = "^0.9.6"
pytest-xdist = "^3.0.2"

[tool.isort]
line_length = 79
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
include = '\.py$'
line-length = 79

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pytest.ini_options]
addopts = "-v -s --durations=0"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "project.settings.local"

# logの設定（ローカル環境）
[local.logging]
version = 1
[local.logging.formatters.simple]
format = "[%(levelname)s] %(name)s %(asctime)s - %(message)s "
[local.logging.handlers.consoleHandler]
class = "logging.StreamHandler"
formatter = "simple"
level = "DEBUG"
stream = "ext://sys.stdout"
[local.logging.handlers.applicationHandler]
class = "logging.handlers.TimedRotatingFileHandler"
filename = "./output/application.log"
formatter = "simple"
level = "INFO"
when = "D"
[local.logging.handlers.errorHandler]
class = "logging.handlers.TimedRotatingFileHandler"
filename = "./output/emergency.log"
formatter = "simple"
level = "ERROR"
when = "D"
[local.logging.loggers.console]
handlers = ["consoleHandler"]
level = "DEBUG"
propagate = "no"
[local.logging.loggers.application]
handlers = ["consoleHandler", "applicationHandler"]
level = "DEBUG"
propagate = "no"
[local.logging.loggers.emergency]
handlers = ["consoleHandler", "errorHandler"]
level = "DEBUG"
propagate = "no"

# logの設定（dev環境）
[dev.logging]
version = 1
[dev.logging.formatters.simple]
format = "[%(levelname)s] %(name)s - %(message)s "
[dev.logging.handlers.consoleHandler]
class = "logging.StreamHandler"
formatter = "simple"
level = "DEBUG"
stream = "ext://sys.stdout"
[dev.logging.loggers.console]
handlers = ["consoleHandler"]
level = "DEBUG"
propagate = "no"
[dev.logging.loggers.application]
handlers = ["consoleHandler"]
level = "DEBUG"
propagate = "no"
[dev.logging.loggers.emergency]
handlers = ["consoleHandler"]
level = "DEBUG"
propagate = "no"
